fix code check

+++ test/fuzztest/player_fuzztest/playerstub_fuzzer/playerstub_fuzzer.cpp
@@ -20,7 +20,7 @@ namespace OHOS {
 namespace Media {
 bool FuzzPlayerStub(uint8_t *data, size_t size)
 {
-    if (data == nullptr || size < sizeof(int32_t)) {
+    if (data == nullptr || size < sizeof(int64_t)) {
         return true;
     }
     constexpr int32_t codeNum = 26;
